# docs/swagger.yaml
definitions:
  api.Airport:
    properties:
      airport_use:
        type: string
      beacon:
        type: boolean
      city:
        type: string
      country_code:
        type: string
      customs:
        type: string
      fuel_types:
        items:
          type: string
        type: array
      iata:
        type: string
      icao:
        type: string
      jet_start_unit:
        type: boolean
      latitude:
        type: number
      longest_runway_ft:
        type: integer
      longitude:
        type: number
      magnetic_variation:
        type: number
      name:
        type: string
      oxygen:
        items:
          type: string
        type: array
      repair_types:
        items:
          type: string
        type: array
      state:
        type: string
      timezone:
        type: string
    type: object
  api.ChartDataResponse:
    properties:
      filename:
        type: string
      georef:
        $ref: '#/definitions/mrvtcl.GeoRefStatus'
      has_tcl:
        type: boolean
      height:
        type: integer
      icao:
        type: string
      width:
        type: integer
    type: object
  api.ChartInfo:
    properties:
      category:
        type: string
      chart_type:
        type: string
      date_eff:
        type: string
      filename:
        type: string
      has_tcl:
        type: boolean
      icao:
        type: string
      is_vfr:
        type: boolean
      proc_id:
        type: string
      sheet_id:
        type: string
      type_name:
        type: string
    type: object
  api.ChartList:
    description: Response containing list of charts for an ICAO
    properties:
      charts:
        items:
          $ref: '#/definitions/api.ChartInfo'
        type: array
      icao:
        type: string
      total:
        type: integer
    type: object
  api.ChartType:
    properties:
      category:
        type: string
      code:
        type: string
      type:
        type: string
    type: object
  api.ChartTypesResponse:
    properties:
      total:
        type: integer
      types:
        items:
          $ref: '#/definitions/api.ChartType'
        type: array
    type: object
  mrvtcl.BatchCoordToPixelRequest:
    properties:
      points:
        items:
          $ref: '#/definitions/mrvtcl.CoordToPixelRequest'
        type: array
    type: object
  mrvtcl.BatchCoordToPixelResponse:
    properties:
      points:
        items:
          $ref: '#/definitions/mrvtcl.CoordToPixelResponse'
        type: array
    type: object
  mrvtcl.BatchPixelToCoordRequest:
    properties:
      points:
        items:
          $ref: '#/definitions/mrvtcl.PixelToCoordRequest'
        type: array
    type: object
  mrvtcl.BatchPixelToCoordResponse:
    properties:
      points:
        items:
          $ref: '#/definitions/mrvtcl.PixelToCoordResponse'
        type: array
    type: object
  mrvtcl.ChartBounds:
    properties:
      bottom:
        format: int32
        type: integer
      height:
        format: int32
        type: integer
      left:
        format: int32
        type: integer
      right:
        format: int32
        type: integer
      top:
        format: int32
        type: integer
      width:
        format: int32
        type: integer
    type: object
  mrvtcl.CoordToPixelRequest:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
  mrvtcl.CoordToPixelResponse:
    properties:
      error:
        type: string
      x:
        type: integer
      "y":
        type: integer
    type: object
  mrvtcl.GeoRefStatus:
    properties:
      bounds:
        $ref: '#/definitions/mrvtcl.ChartBounds'
      georeferenced:
        type: boolean
    type: object
  mrvtcl.PixelToCoordRequest:
    properties:
      x:
        type: integer
      "y":
        type: integer
    type: object
  mrvtcl.PixelToCoordResponse:
    properties:
      error:
        type: string
      latitude:
        type: number
      longitude:
        type: number
    type: object
info:
  contact: {}
paths:
  /api/v1/airports/{icao}:
    get:
      consumes:
      - application/json
      description: Returns airport details for a given ICAO code
      parameters:
      - description: ICAO airport code (e.g., KJFK, EGLL)
        in: path
        name: icao
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Airport'
      summary: Get airport by ICAO
      tags:
      - airports
  /api/v1/chart-types:
    get:
      consumes:
      - application/json
      description: Returns all available chart types from ctypes.dbf with codes, categories,
        and descriptions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ChartTypesResponse'
      summary: List all chart types
      tags:
      - charts
  /api/v1/charts/{icao}:
    get:
      consumes:
      - application/json
      description: Returns all charts for a given ICAO airport. Can be filtered by
        type (code or name) and search query.
      parameters:
      - description: ICAO airport code (e.g., KJFK, EGLL)
        in: path
        name: icao
        required: true
        type: string
      - description: Chart type - can be code (1L, AP) or name (RNAV, ILS, AIRPORT).
          Looks up in ctypes.dbf
        in: query
        name: type
        type: string
      - description: Search text to filter by PROC_ID (procedure name)
        in: query
        name: search
        type: string
      - description: 'Chart types to include - can be ''vfr'', ''ifr'', or ''vfr,ifr''
          (default: both)'
        in: query
        name: types
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ChartList'
      summary: List charts for ICAO
      tags:
      - charts
  /api/v1/charts/{icao}/data/{filename}:
    get:
      consumes:
      - application/json
      description: Returns chart data including dimensions and georeferencing status
        for a specific chart
      parameters:
      - description: ICAO airport code
        in: path
        name: icao
        required: true
        type: string
      - description: Chart filename (e.g., KJFK225)
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ChartDataResponse'
      summary: Get chart data
      tags:
      - charts
  /api/v1/charts/{icao}/export/{filename}:
    get:
      consumes:
      - application/json
      description: Exports a specific chart to PDF format. Returns the chart as a
        PDF file. Post-processing is enabled by default to remove waypoint overlays;
        use ?no_postprocess=1 to disable.
      parameters:
      - description: ICAO airport code
        in: path
        name: icao
        required: true
        type: string
      - description: Chart filename (e.g., KJFK225)
        in: path
        name: filename
        required: true
        type: string
      - description: 'Set to 1 to disable post-processing (default: 0)'
        in: query
        name: no_postprocess
        type: integer
      produces:
      - application/pdf
      responses:
        "200":
          description: PDF file containing the chart
          schema:
            type: file
      summary: Export chart to PDF
      tags:
      - charts
  /api/v1/charts/{icao}/geo/batch-coord2pixel/{filename}:
    post:
      consumes:
      - application/json
      description: Converts multiple geographic coordinates to pixel coordinates
      parameters:
      - description: ICAO airport code
        in: path
        name: icao
        required: true
        type: string
      - description: Chart filename
        in: path
        name: filename
        required: true
        type: string
      - description: Geographic coordinates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mrvtcl.BatchCoordToPixelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mrvtcl.BatchCoordToPixelResponse'
      summary: Batch convert coordinates to pixels
      tags:
      - georef
  /api/v1/charts/{icao}/geo/batch-pixel2coord/{filename}:
    post:
      consumes:
      - application/json
      description: Converts multiple pixel coordinates to geographic coordinates
      parameters:
      - description: ICAO airport code
        in: path
        name: icao
        required: true
        type: string
      - description: Chart filename
        in: path
        name: filename
        required: true
        type: string
      - description: Pixel coordinates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mrvtcl.BatchPixelToCoordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mrvtcl.BatchPixelToCoordResponse'
      summary: Batch convert pixels to coordinates
      tags:
      - georef
  /api/v1/charts/{icao}/geo/coord2pixel/{filename}:
    post:
      consumes:
      - application/json
      description: Converts geographic coordinates (lat, lon) to chart pixel coordinates
        (x, y)
      parameters:
      - description: ICAO airport code
        in: path
        name: icao
        required: true
        type: string
      - description: Chart filename
        in: path
        name: filename
        required: true
        type: string
      - description: Geographic coordinates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mrvtcl.CoordToPixelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mrvtcl.CoordToPixelResponse'
      summary: Convert coordinates to pixel
      tags:
      - georef
  /api/v1/charts/{icao}/geo/pixel2coord/{filename}:
    post:
      consumes:
      - application/json
      description: Converts chart pixel coordinates (x, y) to geographic coordinates
        (lat, lon)
      parameters:
      - description: ICAO airport code
        in: path
        name: icao
        required: true
        type: string
      - description: Chart filename
        in: path
        name: filename
        required: true
        type: string
      - description: Pixel coordinates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/mrvtcl.PixelToCoordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mrvtcl.PixelToCoordResponse'
      summary: Convert pixel to coordinates
      tags:
      - georef
  /api/v1/charts/{icao}/geo/status/{filename}:
    get:
      consumes:
      - application/json
      description: Returns whether a chart is georeferenced and its pixel bounds
      parameters:
      - description: ICAO airport code
        in: path
        name: icao
        required: true
        type: string
      - description: Chart filename (e.g., KJFK225)
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mrvtcl.GeoRefStatus'
      summary: Get chart georeferencing status
      tags:
      - georef
  /health:
    get:
      consumes:
      - application/json
      description: Returns API health status and version
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - health
swagger: "2.0"